# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sws2apps API",
    "version": "3",
    "description": "HTTP contract for version 3 of the sws2apps API."
  },
  "servers": [
    {
      "url": "/api/v3",
      "description": "Current host"
    }
  ],
...
  "paths": {
...
    "/users/{id}/2fa": {
      "parameters": [
        {
          "$ref": "#/components/parameters/UserIdPath"
        }
      ],
      "get": {
        "operationId": "getUserMfaEnrollment",
        "summary": "Get MFA enrollment details",
        "tags": ["User account"],
        "security": [
          {
            "bearerAuth": [],
            "visitorCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AppClientHeader"
          },
          {
            "$ref": "#/components/parameters/AppVersionHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "The account's MFA enrollment details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaEnrollment"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/MfaRequired"
          },
          "403": {
            "$ref": "#/components/responses/AuthenticatedRequestForbidden"
          },
          "404": {
            "$ref": "#/components/responses/AccountContextNotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServerNotReady"
          }
        }
      }
    },
    "/users/{id}/2fa/disable": {
      "parameters": [
        {
          "$ref": "#/components/parameters/UserIdPath"
        }
      ],
      "get": {
        "operationId": "disableUserMfa",
        "summary": "Disable MFA and revoke active sessions",
        "tags": ["User account"],
        "security": [
          {
            "bearerAuth": [],
            "visitorCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AppClientHeader"
          },
          {
            "$ref": "#/components/parameters/AppVersionHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "MFA was disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "MFA_DISABLED"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/MfaRequired"
          },
          "403": {
            "$ref": "#/components/responses/AuthenticatedRequestForbidden"
          },
          "404": {
            "$ref": "#/components/responses/AccountContextNotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServerNotReady"
          }
        }
      }
    },
    "/users/{id}/sessions": {
      "parameters": [
        {
          "$ref": "#/components/parameters/UserIdPath"
        }
      ],
      "get": {
        "operationId": "getUserSessions",
        "summary": "List active account sessions",
        "tags": ["User account"],
        "security": [
          {
            "bearerAuth": [],
            "visitorCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AppClientHeader"
          },
          {
            "$ref": "#/components/parameters/AppVersionHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Safe projections of the account's active sessions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSessionList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/MfaRequired"
          },
          "403": {
            "$ref": "#/components/responses/AuthenticatedRequestForbidden"
          },
          "404": {
            "$ref": "#/components/responses/AccountContextNotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServerNotReady"
          }
        }
      },
      "delete": {
        "operationId": "revokeUserSession",
        "summary": "Revoke an account session",
        "tags": ["User account"],
        "security": [
          {
            "bearerAuth": [],
            "visitorCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AppClientHeader"
          },
          {
            "$ref": "#/components/parameters/AppVersionHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRevocationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The remaining active account sessions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSessionList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/MfaRequired"
          },
          "403": {
            "$ref": "#/components/responses/AuthenticatedRequestForbidden"
          },
          "404": {
            "$ref": "#/components/responses/AccountContextNotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServerNotReady"
          }
        }
      }
    },
    "/users/{id}/erase": {
      "parameters": [
        {
          "$ref": "#/components/parameters/UserIdPath"
        }
      ],
      "delete": {
        "operationId": "deleteUserAccount",
        "summary": "Permanently delete the current account",
        "tags": ["User account"],
        "security": [
          {
            "bearerAuth": [],
            "visitorCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AppClientHeader"
          },
          {
            "$ref": "#/components/parameters/AppVersionHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "The account was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "ACCOUNT_DELETED"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/MfaRequired"
          },
          "403": {
            "$ref": "#/components/responses/AuthenticatedRequestForbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServerNotReady"
          }
        }
      }
    },
...
    "/congregations/admin/{id}/users": {
      "parameters": [{ "$ref": "#/components/parameters/CongregationIdPath" }],
      "get": {
        "operationId": "getCongregationAdministrationMembers",
        "summary": "Get congregation members",
        "tags": ["Congregation administration"],
        "security": [{ "bearerAuth": [], "visitorCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AppClientHeader" }, { "$ref": "#/components/parameters/AppVersionHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/CongregationAdministrationResult" }, "400": { "$ref": "#/components/responses/BadRequest" }, "403": { "$ref": "#/components/responses/CongregationAdministrationForbidden" }, "404": { "$ref": "#/components/responses/CongregationAdministrationNotFound" }, "429": { "$ref": "#/components/responses/TooManyRequests" }, "500": { "$ref": "#/components/responses/InternalError" }, "503": { "$ref": "#/components/responses/ServerNotReady" } }
      },
      "post": {
        "operationId": "addCongregationMember",
        "summary": "Add an existing account to the congregation",
        "tags": ["Congregation administration"],
        "security": [{ "bearerAuth": [], "visitorCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AppClientHeader" }, { "$ref": "#/components/parameters/AppVersionHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CongregationMemberCreateRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/CongregationAdministrationResult" }, "400": { "$ref": "#/components/responses/BadRequest" }, "403": { "$ref": "#/components/responses/CongregationAdministrationForbidden" }, "404": { "$ref": "#/components/responses/CongregationAdministrationNotFound" }, "429": { "$ref": "#/components/responses/TooManyRequests" }, "500": { "$ref": "#/components/responses/InternalError" }, "503": { "$ref": "#/components/responses/ServerNotReady" } }
      }
    },
...
    "/congregations/admin/{id}/users/{user}": {
      "parameters": [{ "$ref": "#/components/parameters/CongregationIdPath" }, { "$ref": "#/components/parameters/CongregationMemberIdPath" }],
      "patch": {
        "operationId": "updateCongregationMember",
        "summary": "Update congregation member details",
        "tags": ["Congregation administration"],
        "security": [{ "bearerAuth": [], "visitorCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AppClientHeader" }, { "$ref": "#/components/parameters/AppVersionHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CongregationMemberUpdateRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/CongregationAdministrationResult" }, "400": { "$ref": "#/components/responses/BadRequest" }, "403": { "$ref": "#/components/responses/CongregationAdministrationForbidden" }, "404": { "$ref": "#/components/responses/CongregationAdministrationNotFound" }, "429": { "$ref": "#/components/responses/TooManyRequests" }, "500": { "$ref": "#/components/responses/InternalError" }, "503": { "$ref": "#/components/responses/ServerNotReady" } }
      },
      "delete": {
        "operationId": "removeCongregationMember",
        "summary": "Remove a member from the congregation",
        "tags": ["Congregation administration"],
        "security": [{ "bearerAuth": [], "visitorCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AppClientHeader" }, { "$ref": "#/components/parameters/AppVersionHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/CongregationAdministrationResult" }, "400": { "$ref": "#/components/responses/BadRequest" }, "403": { "$ref": "#/components/responses/CongregationAdministrationForbidden" }, "404": { "$ref": "#/components/responses/CongregationAdministrationNotFound" }, "429": { "$ref": "#/components/responses/TooManyRequests" }, "500": { "$ref": "#/components/responses/InternalError" }, "503": { "$ref": "#/components/responses/ServerNotReady" } }
      }
    },
...
    "/admin/users": {
      "get": {
        "operationId": "getAllAdministrationUsers",
        "summary": "Get all user accounts",
        "tags": ["Global administration"],
        "security": [{ "bearerAuth": [], "visitorCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AppClientHeader" }, { "$ref": "#/components/parameters/AppVersionHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/AdministrationResult" }, "400": { "$ref": "#/components/responses/BadRequest" }, "403": { "$ref": "#/components/responses/AdministrationForbidden" }, "429": { "$ref": "#/components/responses/TooManyRequests" }, "500": { "$ref": "#/components/responses/InternalError" }, "503": { "$ref": "#/components/responses/ServerNotReady" } }
      }
    },
    "/admin/users/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/AdministrationUserIdPath" }],
      "patch": {
        "operationId": "updateAdministrationUser",
        "summary": "Update user account details",
        "tags": ["Global administration"],
        "security": [{ "bearerAuth": [], "visitorCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AppClientHeader" }, { "$ref": "#/components/parameters/AppVersionHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdministrationUserUpdateRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/AdministrationResult" }, "400": { "$ref": "#/components/responses/BadRequest" }, "403": { "$ref": "#/components/responses/AdministrationForbidden" }, "404": { "$ref": "#/components/responses/AdministrationResourceNotFound" }, "429": { "$ref": "#/components/responses/TooManyRequests" }, "500": { "$ref": "#/components/responses/InternalError" }, "503": { "$ref": "#/components/responses/ServerNotReady" } }
      },
      "delete": {
        "operationId": "deleteAdministrationUser",
        "summary": "Delete a user account",
        "tags": ["Global administration"],
        "security": [{ "bearerAuth": [], "visitorCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AppClientHeader" }, { "$ref": "#/components/parameters/AppVersionHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/AdministrationResult" }, "400": { "$ref": "#/components/responses/BadRequest" }, "403": { "$ref": "#/components/responses/AdministrationForbidden" }, "404": { "$ref": "#/components/responses/AdministrationResourceNotFound" }, "429": { "$ref": "#/components/responses/TooManyRequests" }, "500": { "$ref": "#/components/responses/InternalError" }, "503": { "$ref": "#/components/responses/ServerNotReady" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "description": "Identity token supplied through the Authorization header."
      },
      "visitorCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "visitorid",
        "description": "Signed, HTTP-only session cookie set after authentication."
      }
    },
    "parameters": {
...
      "AdministrationUserIdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "User account identifier selected by a global administrator.",
        "schema": { "type": "string", "minLength": 1 }
      },
...
      "CongregationMemberIdPath": {
        "name": "user",
        "in": "path",
        "required": true,
        "description": "Congregation member account identifier.",
        "schema": { "type": "string", "minLength": 1 }
      },
      "AppClientHeader": {
        "name": "appclient",
        "in": "header",
        "required": true,
        "description": "Calling sws2apps client identifier.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "AppVersionHeader": {
        "name": "appversion",
        "in": "header",
        "required": true,
        "description": "Numeric dotted version of the calling client.",
        "schema": {
          "type": "string",
          "pattern": "^\\d+(?:\\.\\d+)*$"
        }
      },
...
      "CongregationIdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Congregation identifier.",
        "schema": { "type": "string", "minLength": 1 }
      },
...
      "UserIdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Authenticated user's account identifier. Must match the current account.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      }
    },
    "responses": {
      "AccountContextNotFound": {
        "description": "The account, session, or congregation context no longer exists.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "account": {
                "value": {
                  "message": "ACCOUNT_NOT_FOUND"
                }
              },
              "session": {
                "value": {
                  "message": "SESSION_NOT_FOUND"
                }
              },
              "congregation": {
                "value": {
                  "message": "CONG_NOT_FOUND"
                }
              }
            }
          }
        }
      },
      "AuthenticatedRequestForbidden": {
        "description": "Authentication or ownership checks rejected the request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "loginRequired": {
                "value": {
                  "message": "LOGIN_FIRST"
                }
              },
              "deviceRevoked": {
                "value": {
                  "message": "DEVICE_REVOKED"
                }
              },
              "sessionRevoked": {
                "value": {
                  "message": "SESSION_REVOKED"
                }
              },
              "accountMissing": {
                "value": {
                  "message": "ACCOUNT_NOT_FOUND"
                }
              },
              "wrongAccount": {
                "value": {
                  "message": "error_api_unauthorized-request"
                }
              },
              "invalidMfaToken": {
                "value": {
                  "message": "TOKEN_INVALID"
                }
              }
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Request headers, parameters, or payload failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "routeValidation": {
                "value": {
                  "message": "error_api_bad-request"
                }
              },
              "clientValidation": {
                "value": {
                  "message": "INPUT_INVALID"
                }
              },
              "outdatedClient": {
                "value": {
                  "message": "CLIENT_VERSION_OUTDATED"
                }
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "The request could not be completed because of an internal failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "error_api_internal-error"
            }
          }
        }
      },
...
      "MfaRequired": {
        "description": "The active session must complete MFA verification.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "LOGIN_FIRST"
            }
          }
        }
      },
...
      "AdministrationResult": {
        "description": "The global administration operation completed successfully.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/AdministrationResult" }
          }
        }
      },
      "AdministrationForbidden": {
        "description": "The session is missing, revoked, or does not belong to a global administrator.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "examples": {
              "session": { "value": { "message": "LOGIN_FIRST" } },
              "role": { "value": { "message": "UNAUTHORIZED_ACCESS" } }
            }
          }
        }
      },
      "AdministrationResourceNotFound": {
        "description": "The selected user, congregation, feature flag, or request was not found.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "CongregationAdministrationResult": {
        "description": "The congregation administration operation completed successfully.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/CongregationAdministrationResult" }
          }
        }
      },
      "CongregationAdministrationForbidden": {
        "description": "The session is invalid, lacks a congregation administration role, or cannot access the selected congregation.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "examples": {
              "session": { "value": { "message": "LOGIN_FIRST" } },
              "role": { "value": { "message": "UNAUTHORIZED_ACCESS" } },
              "membership": { "value": { "message": "error_api_unauthorized-request" } },
              "masterKey": { "value": { "message": "error_app_security_invalid-master-key" } }
            }
          }
        }
      },
      "CongregationAdministrationNotFound": {
        "description": "The congregation, member account, or join request was not found.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ServerNotReady": {
        "description": "Application state is still initializing.",
        "headers": {
          "Retry-After": {
            "description": "Suggested retry delay in seconds.",
            "schema": {
              "type": "integer",
              "example": 30
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "SERVER_NOT_READY"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "The request rate limit was exceeded.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "TOO_MANY_REQUESTS"
            }
          }
        }
      },
...
    },
    "schemas": {
      "AdministrationResult": {
        "description": "The current administration projection returned by the operation.",
        "oneOf": [
          { "type": "object", "additionalProperties": true },
          { "$ref": "#/components/schemas/RecordList" }
        ]
      },
...
      "AdministrationUserUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["lastname", "firstname", "email", "roles"],
        "properties": {
          "lastname": { "type": "string" },
          "firstname": { "type": "string" },
          "email": { "type": "string", "format": "email", "maxLength": 254 },
          "roles": { "type": "array", "minItems": 1, "items": { "type": "string" } }
        }
      },
...
      "CongregationAdministrationResult": {
        "description": "A refreshed congregation member or request projection, or a stable success message.",
        "oneOf": [
          { "type": "object", "additionalProperties": true },
          { "$ref": "#/components/schemas/RecordList" }
        ]
      },
      "CongregationRoleList": {
        "type": "array",
        "minItems": 1,
        "items": {
          "type": "string",
          "enum": ["admin", "coordinator", "secretary", "service_overseer", "field_service_group_overseer", "midweek_schedule", "weekend_schedule", "public_talk_schedule", "attendance_tracking", "publisher", "view_schedules", "elder", "group_overseers", "language_group_overseers", "duties_schedule"]
        }
      },
...
      "CongregationMemberCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["user_firstname", "user_lastname", "user_id", "cong_role", "cong_person_uid"],
        "properties": {
          "user_firstname": { "type": "string" },
          "user_lastname": { "type": "string" },
          "user_id": { "type": "string", "minLength": 1 },
          "cong_role": { "$ref": "#/components/schemas/CongregationRoleList" },
          "cong_person_uid": { "type": "string" }
        }
      },
      "CongregationMemberUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["cong_role", "cong_person_uid", "cong_person_delegates", "first_name", "last_name"],
        "properties": {
          "cong_role": { "$ref": "#/components/schemas/CongregationRoleList" },
          "cong_person_uid": { "type": "string" },
          "cong_person_delegates": { "type": "array", "items": { "type": "string" } },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "user_secret_code": { "type": "string", "description": "Optional replacement Pocket invitation code." }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "MfaEnrollment": {
        "type": "object",
        "additionalProperties": false,
        "required": ["secret", "qrCode"],
        "properties": {
          "secret": {
            "type": "string"
          },
          "qrCode": {
            "type": "string",
            "description": "An otpauth provisioning URI suitable for a QR code."
          },
          "mfaEnabled": {
            "type": "boolean"
          },
          "MFA_CODE": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "description": "Development-only token; omitted in production."
          }
        }
      },
...
      "RecordList": {
        "type": "array",
        "items": {
          "type": "object",
          "additionalProperties": true
        }
      },
      "SessionRevocationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["identifier"],
        "properties": {
          "identifier": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          }
        }
      },
      "SuccessMessage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "UserSession": {
        "type": "object",
        "additionalProperties": false,
        "required": ["identifier", "isSelf", "ip", "country_name", "device", "last_seen"],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "isSelf": {
            "type": "boolean"
          },
          "ip": {
            "type": "string"
          },
          "country_name": {
            "type": "string"
          },
          "device": {
            "type": "object",
            "additionalProperties": false,
            "required": ["browserName", "os", "isMobile"],
            "properties": {
              "browserName": {
                "type": "string"
              },
              "os": {
                "type": "string"
              },
              "isMobile": {
                "type": "boolean"
              }
            }
          },
          "last_seen": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserSessionList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/UserSession"
        }
      },
...
    }
  }
}